fixed bug: can message only 1 uid
earlier, msg.Sent was set to true for 1st UID,
so for the second UID, this value was read true,
and the sender wasn't allowed to send anything else
to anyone else.

Now, sent.${UID} is set to true.

Since it stores the specific UID, it prevents the sender
from spamming a specific user, but allows sending message
to anothwr user (once).